Python/requirements.txt
azure-identity
msgraph-sdk
pymongo
python-dotenv
